# docs/v1/swaggerv1.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Hanz Car Rentals API",
		"description": "This API is made for Hanz Car Rentals as a school project.",
		"version": "1.2.0",
		"contact": {
			"name": "Hanz",
			"email": "hanz.car.rentals@wolfsoft.solutions",
			"url": "https://hanzcar.rentals/support"
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"responses": {
			"UnauthorizedError": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"BadRequestError": {
				"description": "Bad request",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"NotFoundError": {
				"description": "Not found",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"ServerError": {
				"description": "Server error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"status": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					}
				}
			}
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
	"servers": [
		{
			"url": "https://api.hanzcar.rentals/api/v1/",
			"description": "Production server"
		},
		{
			"url": "http://localhost:55600/api/v1/",
			"description": "Development server"
		}
	],
	"paths": {
...
		"/cars": {
			"get": {
				"summary": "Get all cars",
				"tags": [
					"Cars"
				],
				"description": "This endpoint returns all cars",
				"security": [],
				"responses": {
					"200": {
						"description": "A list of cars",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"brand": {
												"type": "string"
											},
											"model": {
												"type": "string"
											},
											"year": {
												"type": "integer"
											},
											"color": {
												"type": "string"
											},
											"price": {
												"type": "integer"
											},
											"rented_by": {
												"type": "string"
											},
											"licence_plate": {
												"type": "string"
											},
											"seats": {
												"type": "integer"
											},
											"space": {
												"type": "string"
											},
											"transmission": {
												"type": "string"
											},
											"fuel": {
												"type": "string"
											},
											"doors": {
												"type": "integer"
											},
											"towing_weight": {
												"type": "string"
											},
											"maximum_gross_weight": {
												"type": "string"
											},
											"location": {
												"type": "string"
											},
											"image": {
												"type": "string"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/cars/car/{id}": {
			"get": {
				"summary": "Get a car by ID",
				"tags": [
					"Cars"
				],
				"description": "This endpoint returns a car by its ID",
				"security": [],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "A car object",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "integer"
										},
										"brand": {
											"type": "string"
										},
										"model": {
											"type": "string"
										},
										"year": {
											"type": "integer"
										},
										"color": {
											"type": "string"
										},
										"price": {
											"type": "integer"
										},
										"rented_by": {
											"type": "string"
										},
										"licence_plate": {
											"type": "string"
										},
										"seats": {
											"type": "integer"
										},
										"space": {
											"type": "string"
										},
										"transmission": {
											"type": "string"
										},
										"fuel": {
											"type": "string"
										},
										"doors": {
											"type": "integer"
										},
										"towing_weight": {
											"type": "string"
										},
										"maximum_gross_weight": {
											"type": "string"
										},
										"location": {
											"type": "string"
										},
										"image": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
...
		"/users": {
			"get": {
				"summary": "Get all users",
				"tags": [
					"Users"
				],
				"description": "This endpoint returns all users",
				"responses": {
					"200": {
						"description": "A list of users",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"first_name": {
												"type": "string"
											},
											"last_name": {
												"type": "string"
											},
											"email": {
												"type": "string"
											},
											"email_verified": {
												"type": "integer"
											},
											"admin": {
												"type": "integer"
											},
											"times_rented": {
												"type": "integer"
											},
											"currently_renting": {
												"type": "integer"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/users/user/{id}": {
			"get": {
				"summary": "Get a user by ID",
				"tags": [
					"Users"
				],
				"description": "This endpoint returns a user by its ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "A user object",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "integer"
										},
										"email": {
											"type": "string"
										},
										"email_verified": {
											"type": "integer"
										},
										"first_name": {
											"type": "string"
										},
										"last_name": {
											"type": "string"
										},
										"admin": {
											"type": "integer"
										},
										"times_rented": {
											"type": "integer"
										},
										"currently_renting": {
											"type": "integer"
										}
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/users/add": {
			"post": {
				"summary": "Add a user",
				"tags": [
					"Users"
				],
				"description": "This endpoint adds a user",
				"security": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"fname": {
										"type": "string"
									},
									"lname": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User added"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/users/update/{id}": {
			"put": {
				"summary": "Update a user",
				"tags": [
					"Users"
				],
				"description": "This endpoint updates a user",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"fname": {
										"type": "string"
									},
									"lname": {
										"type": "string"
									},
									"admin": {
										"type": "integer"
									}
								}
							}
						}
					},
					"responses": {
						"200": {
							"description": "User updated"
						},
						"401": {
							"$ref": "#/components/responses/UnauthorizedError"
						},
						"403": {
							"$ref": "#/components/responses/BadRequestError"
						},
						"404": {
							"$ref": "#/components/responses/NotFoundError"
						},
						"500": {
							"$ref": "#/components/responses/ServerError"
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"summary": "Login",
				"tags": [
					"Users"
				],
				"description": "This endpoint logs in a user",
				"security": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"status": "200",
						"message": "Logged in successfully",
						"token": "UserTokenHere",
						"user": {
							"id": 12,
							"admin": 0
						}
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/users/reset_password_email": {
			"post": {
				"summary": "Reset password",
				"tags": [
					"Users"
				],
				"description": "This endpoint sends an email with a token for the password reset function endpoint.",
				"security": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Password reset"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
		"/users/reset_password": {
			"post": {
				"summary": "Reset password",
				"tags": [
					"Users"
				],
				"description": "This endpoint resets a user's password",
				"security": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"password1": {
										"type": "string"
									},
									"password2": {
										"type": "string"
									},
									"email_token": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Password reset"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"403": {
						"$ref": "#/components/responses/BadRequestError"
					},
					"404": {
						"$ref": "#/components/responses/NotFoundError"
					},
					"500": {
						"$ref": "#/components/responses/ServerError"
					}
				}
			}
		},
...
	}
}